+++ packages/common/src/types/field.ts
@@ -37,6 +37,7 @@ export enum NumberSeparator {
 type CompactConfig = {
     compact: Compact;
     alias: Array<typeof CompactAlias[number]>;
+    orderOfMagnitude: number;
     convertFn: (value: number) => number;
     label: string;
     suffix: string;
@@ -48,6 +49,7 @@ export const CompactConfigMap: Record<Compact, CompactConfig> = {
     [Compact.THOUSANDS]: {
         compact: Compact.THOUSANDS,
         alias: ['K', 'thousand'],
+        orderOfMagnitude: 3,
         convertFn: (value: number) => value / 1000,
         label: 'thousands (K)',
         suffix: 'K',
@@ -55,6 +57,7 @@ export const CompactConfigMap: Record<Compact, CompactConfig> = {
     [Compact.MILLIONS]: {
         compact: Compact.MILLIONS,
         alias: ['M', 'million'],
+        orderOfMagnitude: 6,
         convertFn: (value: number) => value / 1000000,
         label: 'millions (M)',
         suffix: 'M',
@@ -62,6 +65,7 @@ export const CompactConfigMap: Record<Compact, CompactConfig> = {
     [Compact.BILLIONS]: {
         compact: Compact.BILLIONS,
         alias: ['B', 'billion'],
+        orderOfMagnitude: 9,
         convertFn: (value: number) => value / 1000000000,
         label: 'billions (B)',
         suffix: 'B',
@@ -69,6 +73,7 @@ export const CompactConfigMap: Record<Compact, CompactConfig> = {
     [Compact.TRILLIONS]: {
         compact: Compact.TRILLIONS,
         alias: ['T', 'trillion'],
+        orderOfMagnitude: 12,
         convertFn: (value: number) => value / 1000000000000,
         label: 'trillions (T)',
         suffix: 'T',
@@ -245,9 +250,11 @@ export interface Field {
     description?: string;
     source?: Source | undefined;
     hidden: boolean;
+    // @deprecated Use format expression instead
     compact?: CompactOrAlias;
+    // @deprecated Use format expression instead
     round?: number;
-    format?: Format;
+    format?: Format | string; // Format type is deprecated, use format expression(string) instead
     /**
      * @deprecated Use groups property instead.
      */