+++ packages/frontend/src/components/Explorer/ExploreTree/TableTree/Tree/TreeSingleNodeActions.tsx
@@ -236,7 +236,7 @@ const TreeSingleNodeActions: FC<Props> = ({
                                     ) => {
                                         e.stopPropagation();
                                         toggleAdditionalMetricWriteBackModal({
-                                            item,
+                                            items: [item],
                                         });
                                     }}
                                 >